Rename gateway_endpoints create_request.
