+++ crates/papyrus_load_test/src/lib.rs
@@ -1,4 +1,4 @@
-pub mod gateway_endpoints;
+pub mod create_request;
 pub mod scenarios;
 pub mod transactions;
